tests/wpt/tests/webauthn/credblob-not-supported.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>credBlob extension tests</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
"use strict";

const blobu8 = new Uint8Array(16);
window.crypto.getRandomValues(blobu8);
const blob = blobu8.buffer;

virtualAuthenticatorPromiseTest(async t => {
  const cred = await createCredential({
    options: {
      publicKey: {
        extensions: {
          credBlob: blob,
        },
      },
    },
  });

  const createExtensions = cred.getClientExtensionResults();
  assert_own_property(createExtensions, "credBlob");
  assert_equals(createExtensions.credBlob, false, "extension supported at create time");
}, {
  protocol: "ctap2_1",
  extensions: [],
}, "creation requesting credBlob without authenticator support");
</script>
